Refactored code - no new features added
